+++ src/wing/CCPACSWingSegments.cpp
@@ -223,6 +223,9 @@ CCPACSWingSegment& CCPACSWingSegments::SplitSegment(const std::string& segmentUI
     segment.SetToElementUID(splitterElement.GetUID());
 
     Invalidate();
+    // Reordering is necessary: After splitting a segment and adding another one, the order is not correct
+    // Without reordering, this results in a wrong shape visible in the TiGLViewer
+    ReorderSegments();
 
     return additionalSegment;
 }